Plant Species Classifier Using ANN
At Rajarata University of Sri Lanka, there are over a thousand plant families, many of which are not well-identified or maintained. To address this, we are creating a classifier to facilitate easy species identification, which may be integrated into a mobile app or web application in future for widespread use. As an initial step, we focus on three species within the Bougainvillea genus: Bougainvillea buttiana, Bougainvillea glabra, and Bougainvillea spectabilis.

Bougainvillea buttiana is known for its vibrant bracts and resilience in various climates, Bougainvillea glabra, often called paper flower, is distinguished by its smooth leaves and thorny stems, and Bougainvillea spectabilis, notable for its larger and more robust flowers. This project not only enhances botanical research but also contributes to environmental conservation efforts by promoting awareness and proper management of plant species.